Move candles sync-schema to separate file
